pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-glamor-allure"
dynamic = ["version"]
description = "Extends allure-pytest functionality"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
    'Framework :: Pytest',
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'Natural Language :: Russian',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3.14',
    'Topic :: Software Development :: Testing',
    'Topic :: Software Development :: Quality Assurance',
]

dependencies = [
    "allure-pytest<=2.15.3",
    "allure-python-commons",
    "pytest<=9.0.2",
    "attrs<=25.4.0",
]
authors = [{ name = "Denis Alexeev", email = "herr.alekseev.denis@gmail.com" }]
maintainers = [
    { name = "Denis Alexeev", email = "herr.alekseev.denis@gmail.com" },
]


[project.entry-points.pytest11]
pytest_glamor_allure = "pytest_glamor_allure.plugin"

[project.urls]
Repository = "https://github.com/Denis-Alexeev/pytest-glamor-allure"
Issues = "https://github.com/Denis-Alexeev/pytest-glamor-allure/issues"
Documentation = "https://denis-alexeev.github.io/pytest-glamor-allure/"
Homepage = "https://github.com/Denis-Alexeev/pytest-glamor-allure"

[tool.setuptools]
packages = ["pytest_glamor_allure", "glamor"]
py-modules = ["pitest"]
include-package-data = true

[tool.setuptools_scm]
version_file = "pytest_glamor_allure/_version.py"

[tool.ruff]
exclude = ["venv*", ".tox", ".eggs", "pytest_glamor_allure/_version.py"]
line-length = 79
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "Q000",   # Single quotes found but double quotes preferred
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D107",   # Missing docstring in __init__
    "ANN204", # Missing return type annotation for special method __init__
    "ANN202", # Missing return type annotation for private function
    "SLF001", # Private member accessed
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]
"pitest.py" = ["F401", "F403", "F405"]
"tests/**" = [
    "S101",    # Use of `assert` detected
    "PLR2004", # Magic value used in comparison
    "ANN201",  # Missing return type annotation for public function
    "ANN001",  # Missing type annotation for function argument
    "ANN002",  # Missing type annotation for `*args`
    "ANN003",  # Missing type annotation for `**kwargs`
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "ARG002",  # Unused method argument
    "PLR0913", # Too many arguments in function definition
]

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.lint.isort]
from-first = true
combine-as-imports = true

[tool.coverage.run]
source = ["pytest_glamor_allure", "glamor", "glamor.patches", "pitest"]
branch = true
parallel = true
patch = ["subprocess"]
omit = ["tests/*", "pytest_glamor_allure/_version.py"]

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_also = ["if TYPE_CHECKING:"]
